A bit of reorganizing :)
