+++ Protocol/Protocol.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{69CFF08E-96BF-4839-A570-3D48BB023CF6}</ProjectGuid>
+    <ProjectGuid>{29498D6F-071D-43E5-A0F7-4369F78A4B7C}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Miner</RootNamespace>
-    <AssemblyName>Miner</AssemblyName>
+    <RootNamespace>Protocol</RootNamespace>
+    <AssemblyName>Protocol</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -41,23 +41,25 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Miner.cs" />
-    <Compile Include="MinerManager.cs" />
+    <Compile Include="Algorithm.cs" />
+    <Compile Include="BlockPayload.cs" />
+    <Compile Include="DifficultyPayload.cs" />
+    <Compile Include="IntegerPayload.cs" />
+    <Compile Include="InvEntryPayload.cs" />
+    <Compile Include="InvPayload.cs" />
+    <Compile Include="IPAddressPayload.cs" />
+    <Compile Include="MessagePayload.cs" />
+    <Compile Include="Payload.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Blockchain\Blockchain.csproj">
-      <Project>{62bbd29d-7847-4097-a84f-bf1da96dcdf4}</Project>
-      <Name>Blockchain</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\Connection\Connection.csproj">
-      <Project>{3905a3ea-5cc9-4fc8-ba8c-df4b6f5baa3e}</Project>
-      <Name>Connection</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\Protocol\Protocol.csproj">
-      <Project>{29498d6f-071d-43e5-a0f7-4369f78a4b7c}</Project>
-      <Name>Protocol</Name>
-    </ProjectReference>
+    <Compile Include="ProtocolConfiguration.cs" />
+    <Compile Include="StringPayload.cs" />
+    <Compile Include="TransactionPayload.cs" />
+    <Compile Include="TxInputPayload.cs" />
+    <Compile Include="TxOutPointPayload.cs" />
+    <Compile Include="TxOutputPayload.cs" />
+    <Compile Include="UnknownPayload.cs" />
+    <Compile Include="VerAckPayload.cs" />
+    <Compile Include="VersionPayload.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
