+++ ProtocolTest/Properties/AssemblyInfo.cs
@@ -2,18 +2,18 @@ using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("ConnectionTest")]
+[assembly: AssemblyTitle("ProtocolTest")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("ConnectionTest")]
+[assembly: AssemblyProduct("ProtocolTest")]
 [assembly: AssemblyCopyright("Copyright ©  2018")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
 [assembly: ComVisible(false)]
 
-[assembly: Guid("cdb9167e-8bcc-4b74-b837-3b8d9f9f5343")]
+[assembly: Guid("2f33e7b9-90a4-4502-8849-67ba2a1ca0b0")]
 
 // [assembly: AssemblyVersion("1.0.*")]
 [assembly: AssemblyVersion("1.0.0.0")]
