+++ ProtocolTest/ProtocolTest.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{CDB9167E-8BCC-4B74-B837-3B8D9F9F5343}</ProjectGuid>
+    <ProjectGuid>{2F33E7B9-90A4-4502-8849-67BA2A1CA0B0}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>PrimeNetworkTest</RootNamespace>
-    <AssemblyName>PrimeNetworkTest</AssemblyName>
+    <RootNamespace>ProtocolTest</RootNamespace>
+    <AssemblyName>ProtocolTest</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{3AC096D0-A1C2-E12C-1390-A8335801FDAB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
@@ -49,16 +49,24 @@
     <Reference Include="System.Numerics" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="PayloadTest.cs" />
+    <Compile Include="TestAlgorithm.cs" />
+    <Compile Include="TestBlockPayload.cs" />
+    <Compile Include="TestDifficultyPayload.cs" />
+    <Compile Include="TestIntegerPayload.cs" />
+    <Compile Include="TestIPAddressPayload.cs" />
+    <Compile Include="TestMessagePayload.cs" />
+    <Compile Include="TestPayload.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TestStringPayload.cs" />
+    <Compile Include="TestVersionPayload.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Connection\Connection.csproj">
-      <Project>{3905a3ea-5cc9-4fc8-ba8c-df4b6f5baa3e}</Project>
-      <Name>Connection</Name>
+    <ProjectReference Include="..\Protocol\Protocol.csproj">
+      <Project>{29498d6f-071d-43e5-a0f7-4369f78a4b7c}</Project>
+      <Name>Protocol</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(VSToolsPath)\TeamTest\Microsoft.TestTools.targets" Condition="Exists('$(VSToolsPath)\TeamTest\Microsoft.TestTools.targets')" />
