+++ WinPrimecoin/WinPrimecoin.csproj
@@ -107,6 +107,10 @@
       <Project>{69cff08e-96bf-4839-a570-3d48bb023cf6}</Project>
       <Name>Miner</Name>
     </ProjectReference>
+    <ProjectReference Include="..\Protocol\Protocol.csproj">
+      <Project>{29498d6f-071d-43e5-a0f7-4369f78a4b7c}</Project>
+      <Name>Protocol</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
